Rename folder section.ts to mediaQueries.ts to change it to a proper name
